refactor: 트랜잭션 공통 로직 분리
