+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -2,22 +2,16 @@ package com.techcourse.service;
 
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
-import org.springframework.jdbc.datasource.DataSourceUtils;
-import org.springframework.jdbc.exception.DataAccessException;
-import org.springframework.jdbc.exception.RollbackFailException;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
+import com.techcourse.support.transaction.TransactionExecutor;
 
 public class TxUserService {
 
     private final UserService userService;
-    private final DataSource dataSource;
+    private final TransactionExecutor transactionExecutor;
 
     public TxUserService(final UserService userService) {
         this.userService = userService;
-        this.dataSource = DataSourceConfig.getInstance();
+        this.transactionExecutor = new TransactionExecutor(DataSourceConfig.getInstance());
     }
 
     public User findById(final long id) {
@@ -25,28 +19,10 @@ public class TxUserService {
     }
 
     public void insert(final User user) {
-        userService.insert(user);
+        transactionExecutor.execute(() -> userService.insert(user));
     }
 
-    public void changePassword(final long id, final String newPassword, final String createBy) throws SQLException {
-        final Connection connection = DataSourceUtils.getConnection(dataSource);
-        connection.setAutoCommit(false);
-        try {
-            userService.changePassword(id, newPassword, createBy);
-            connection.commit();
-        } catch (SQLException e) {
-            rollback(connection);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
-
-    private static void rollback(Connection connection) {
-        try {
-            connection.rollback();
-            throw new DataAccessException("데이터에 접근할 수 없습니다.");
-        } catch (SQLException exception) {
-            throw new RollbackFailException("롤백을 실패했습니다.");
-        }
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        transactionExecutor.execute(() -> userService.changePassword(id, newPassword, createBy));
     }
 }
