+++ app/src/main/java/com/techcourse/support/transaction/ServiceLogicExecutor.java
@@ -1,0 +1,9 @@
+package com.techcourse.support.transaction;
+
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface ServiceLogicExecutor {
+
+    void execute() throws SQLException;
+}
